% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/boot_params.R
\name{boot.params}
\alias{boot.params}
\title{Calculate the Bowes and Greene parameters for a given set of paired Q and P}
\usage{
boot.params(Q, P, Qhi = NULL, N = 50)
}
\arguments{
\item{Q}{the measured water flow rate in cubic metres per second}

\item{P}{the Total Reactive Phostphate - units- TBD}

\item{Qhi}{the high-frequency Q values in cubic metres per second}

\item{N}{the number of bootstrap replications - defaults to 50 for a reasonable size for estimating the mean.}
}
\description{
Calculate the Bowes and Greene parameters for a given set of paired Q and P
}

