% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{Bowes.calcs}
\alias{Bowes.calcs}
\title{Bowes calculations}
\usage{
Bowes.calcs(d1, AltQ = NULL)
}
\arguments{
\item{d1}{a data.frame in a particular format - TBD}

\item{AltQ}{an optional numeric vector containing the flows to be used
for apportionment, nb. in m^3/sec}
}
\value{
a list containing the results
}
\description{
Takes a data.frame and performs a number of calculations related
to the Bowes formal
}
\details{
Does all the Bowes calculations as far as apportionment
nb AltQ is an alternative source of Q - typically High-Frequency Q
when we only have low frequency TRP
when AltQ is not supplied, then the Q values from d1 are used
NB. AltQ is assumed to be a numeric array identical to
  d1$Q, but possibly much much longer!
}

