% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{Bowes.fit}
\alias{Bowes.fit}
\title{Fit the Bowes A,B,C,D parameters to data}
\usage{
Bowes.fit(d1, bf.iMax = 1e+06)
}
\arguments{
\item{d1}{dataframe containing...}

\item{bf.iMax}{maximum number of iterations of XXX defaulting to a million}
}
\description{
Fit the Bowes A,B,C,D parameters to data
}

