% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowes_recalc.R
\name{Bowes.ReCalc}
\alias{Bowes.ReCalc}
\title{Do the recalculations necessary for re-analysing with a HiFrequency dataset.}
\usage{
Bowes.ReCalc(arec, Qsource = NULL)
}
\arguments{
\item{arec}{is a record from the output files which have names like}

\item{Qsource}{is a numeric vector of flow rates in cubic metres per second}
}
\description{
Do the recalculations necessary for re-analysing with a HiFrequency dataset.
}
