% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/calc_params.R
\name{calc.params}
\alias{calc.params}
\title{Calculate the Bowes and Greene parameters for a given set of paired Q and P}
\usage{
calc.params(Q, P, Qhi = NULL)
}
\arguments{
\item{Q}{the measured water flow rate in cubic metres per second}

\item{P}{the Total Reactive Phostphate - units- TBD}

\item{Qhi}{the high-frequency Q values in cubic metres per second}
}
\description{
Calculate the Bowes and Greene parameters for a given set of paired Q and P
}

