% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/greene_calcs.R
\name{Greene.calcs}
\alias{Greene.calcs}
\title{Green calculations as far as apportionment}
\usage{
Greene.calcs(d1, AltQ = NULL)
}
\arguments{
\item{d1}{a dataframe containing XXX}

\item{AltQ}{a numeric vector containing an alternative source of Q information which will be used for XXX}
}
\description{
lets do all the Greene calculations as far as apportionment
}
\details{
nb AltQ is an alternative source of Q - typically High-Frequency Q
when we only have low frequency TRP
when AltQ is not supplied, then the Q values from d1 are used
NB. AltQ is assumed to be a numeric array identical to
  d1$Q, but possibly much much longer!
}

