% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{Greene.fit}
\alias{Greene.fit}
\title{Fit the Greene parameters to data}
\usage{
Greene.fit(d1, iMax = 100)
}
\arguments{
\item{d1}{a data.frame XXX}

\item{iMax}{maximum number of iterations? XXX}
}
\description{
\code{Greene.fit} takes a dataset and attempts to fit a Greene Model, i.e.
$Y = A/X   + B*X     + C*X^2$, i.e. Y = Point + Diffuse + Hysteresis
with   no constraints
}
\details{
Note that sometimes this has problems converging hence
there is a loop and random start points
}

