% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeneqe.R
\name{GreeneQe}
\alias{GreeneQe}
\title{Solve the Greene equation for Qe}
\usage{
GreeneQe(A, B, C, rightedge = 1, sDEBUG = NULL)
}
\arguments{
\item{A}{Greene equation coefficient}

\item{B}{Greene equation coefficient}

\item{C}{Greene equation coefficient}

\item{rightedge}{The flow rate parameter which is expected to be greater than Qe}

\item{sDEBUG}{a parameter not normally used except in debugging, set to be an
integer >5 for maximum debug information printed.}
}
\description{
Qe is when the larger contribution switches from
point to diffuse - there is always point contribution but at high flows it is
outweighed by the diffuse contribution.
\code{GreeneQe} takes a set of three Greene coefficient, A,B,C and solves
the Greene equation for the
$TRP = A/Q   + B*Q     + C*Q^2$, i.e. TRP = Point + Diffuse + Hysteresis
flow rate Q at which Point = Diffuse + Hysteresis
}
\details{
Note that sometimes this has problems converging hence
there is a further parater right edge which limits the range of Q at the
upper end.
}
