% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{label.date.range}
\alias{label.date.range}
\title{take a data.frame and label dates according to given ranges}
\usage{
label.date.range(dates, date.range.df, else.label = "Not Assigned")
}
\arguments{
\item{dates}{vector of dates which are to be classifed}

\item{date.range.df}{data.frame specifying how to classify dates}

\item{else.label}{if a date isn't in a specified range then use this label}
}
\description{
take a data.frame and label dates according to given ranges
}
\details{
start a list of start months
  end a list of end months, nb the list must be the same length as start
  label a list of labels to be attached, nb the list must be the same length as start and end
}

