% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsamp_fixeddays.R
\name{subsamp.fixeddays}
\alias{subsamp.fixeddays}
\title{Subsample one sample each from fixed days of the week}
\usage{
subsamp.fixeddays(adf, sdays = c("Mon", "Wed", "Thu"))
}
\arguments{
\item{adf}{the input data.frame}

\item{sdays, }{the fixed days on which to take the samples,
by default c('Mon','Wed','Thu')}
}
\description{
subsample rows from a data.frame that are
three days per week fixed days, but random time working hours
}
