% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsamp_ndays.R
\name{subsamp.ndays}
\alias{subsamp.ndays}
\title{Choose a subset of days and subsample from each of these days once}
\usage{
subsamp.ndays(adf, ndays = 3)
}
\arguments{
\item{adf}{the input data.frame}

\item{ndays}{the number of days to subsample from 1<= ndays <= 7}
}
\description{
subsample rows from a data.frame that are
only from three randomly chosen working days, note it
is the days (i.e. Mon to Sun) that are randomly chosen.
}
