% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{subsamp.nperday}
\alias{subsamp.nperday}
\title{subsample n samples per day}
\usage{
subsamp.nperday(adf, n = 1)
}
\arguments{
\item{adf}{the input data.frame}

\item{n}{the number of samples to take per day}
}
\description{
subsample rows from a data.frame that are
only n samples per day
}

