% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsamp_nrow.R
\name{subsamp.nrow}
\alias{subsamp.nrow}
\title{subsample from a data.frame}
\usage{
subsamp.nrow(adf, n = 1)
}
\arguments{
\item{adf}{a data.frame to sample from}

\item{n}{number of rows to sample, default 1}
}
\value{
rows from the input data.frame as a data.frame
}
\description{
sub-sample rows randomly from a data.frame
}
