% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{test.label.date.range}
\alias{test.label.date.range}
\title{test label date range}
\usage{
test.label.date.range(N = 40, random.seed = 123)
}
\arguments{
\item{N}{testing number}

\item{random.seed}{just that - a random seed which defaults to 123}
}
\description{
test label date range
}

