% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/week_of_month.R
\name{week.of.month}
\alias{week.of.month}
\title{subsample once from a particular day on the fourth week of the month}
\usage{
week.of.month(adf, which.day = "Mon", which.week = 4,
  only.workinghours = TRUE)
}
\arguments{
\item{adf}{the input data.frame}

\item{which.day}{- 'Mon' by default, three letter day in locale (I presume)}

\item{which.week}{- 4 (th) week of the month by default - an integer}

\item{only.workinghours}{- default TRUE}
}
\description{
subsample rows from a data.frame that are
on a particular day (Monday by default) on the fourth week of the month
and they could be at anytime, working hours by default,
}

