% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_params.R
\name{boot.params}
\alias{boot.params}
\title{Calculate the Bowes and Greene parameters for a given set of paired Q and P}
\usage{
boot.params(Q, P, Qhi = NULL, N = 50, ConstrainBzero = FALSE)
}
\arguments{
\item{Q}{the measured water flow rate in cubic metres per second}

\item{P}{the Total Reactive Phosphate - units- TBD}

\item{Qhi}{the high-frequency Q values in cubic metres per second}

\item{N}{the number of bootstrap replications - defaults to 50 for a reasonable size for estimating the mean.}

\item{ConstrainBzero}{Do we constrain B to be zero - default FALSE}
}
\description{
Calculate the Bowes and Greene parameters for a given set of paired Q and P
}
\details{
Note that the time taken is approximately linear in N
  but not entirely deterministic as the starting positions for
  identifying the parameters in the nonlinear models are chosen at
  random and occasionally you might see an error message "Error in nlsModel(formula, mf, start, wts) :
   singular gradient matrix at initial parameter estimates", which
  can be ignored for now.
  In some cases, particularly when you have a very small TRP data set,
  the `nls()` function can have difficulty in finding the parameters
  and this has been described as finding the minimum point in a valley
  but it turns out that the valley has an ice lake and a lot of time
  can be taken searching for the lowest point on this ice lake until
  eventually it finds a small pit in the ice!
}
