% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowes_fit.R
\name{Bowes.fit}
\alias{Bowes.fit}
\title{Fit the Bowes A,B,C,D parameters to data}
\usage{
Bowes.fit(d1, bf.iMax = 10000, ConstrainBzero = FALSE)
}
\arguments{
\item{d1}{dataframe containing a column "Q" and a column "TRP"}

\item{bf.iMax}{maximum number of iterations of XXX defaulting to a million}

\item{ConstrainBzero}{if TRUE (default), then the Bowes model will constrain
B to be zero instead of allowing it to be in the range 0.01 to 1}
}
\description{
Fit the Bowes A,B,C,D parameters to data
}
