% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_params.R
\name{calc.params}
\alias{calc.params}
\title{Calculate the Bowes and Greene parameters for a given set of paired Q and P}
\usage{
calc.params(Q, P, Qhi = NULL, ConstrainBzero = FALSE)
}
\arguments{
\item{Q}{the measured water flow rate in cubic metres per second}

\item{P}{the Total Reactive Phostphate - units- TBD}

\item{Qhi}{the high-frequency Q values in cubic metres per second}

\item{ConstrainBzero}{if TRUE (default), then the Bowes model will constrain
B to be zero instead of allowing it to be in the range 0.01 to 1}
}
\description{
Calculate the Bowes and Greene parameters for a given set of paired Q and P
}
