% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/greene_fit.R
\name{Greene.fit}
\alias{Greene.fit}
\title{Fit the Greene parameters to data}
\usage{
Greene.fit(d1, iMax = 100, sDEBUG = NULL)
}
\arguments{
\item{d1}{a data.frame XXX}

\item{iMax}{maximum number of iterations? Refer to the Detailed Description.}

\item{sDEBUG}{a parameter not normally used except in debugging, set to be an
integer >5 for maximum debug information printed.}
}
\description{
\code{Greene.fit} takes a dataset and attempts to fit a Greene Model, i.e.
$Y = A/X   + B*X     + C*X^2$, i.e. Y = Point + Diffuse + Hysteresis
with   no constraints
}
\details{
Note that sometimes this has problems converging hence
there is a loop and random start points.

Note too that there will be only iMax iterations of the \code{stats:nls()}
function before the \code{Greene.fit()} function gives up.  However, the
 \code{stats:nls()} itself is parameterized to run 500 iterations, so the
 overall runtime is likely to take some time before it gives up with a
 \code{stop()} message.
}

