% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsamp_nperday.R
\name{subsamp.nperday}
\alias{subsamp.nperday}
\title{subsample n samples per day}
\usage{
subsamp.nperday(adf, n = 1)
}
\arguments{
\item{adf}{the input data.frame}

\item{n}{the number of samples to take per day}
}
\description{
subsample rows from a data.frame that are
only n samples per day
}
