% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_date_range.r
\name{test.label.date.range}
\alias{test.label.date.range}
\title{test label date range}
\usage{
test.label.date.range(N = 40, random.seed = 123)
}
\arguments{
\item{N}{testing number}

\item{random.seed}{just that - a random seed which defaults to 123}
}
\description{
test label date range
}
